+++ code/packages/sdk/src/nodes/image/image.ts
@@ -19,7 +19,6 @@ import {
 import { WeaveImageToolAction } from '@/actions/image-tool/image-tool';
 import { WeaveImageCrop } from './crop';
 import { WEAVE_IMAGE_CROP_END_TYPE, WEAVE_IMAGE_NODE_TYPE } from './constants';
-import type { WeaveNodesSelectionPlugin } from '@/plugins/nodes-selection/nodes-selection';
 import { isEqual } from 'lodash';
 import { WEAVE_STAGE_DEFAULT_MODE } from '../stage/constants';
 import { IMAGE_TOOL_ACTION_NAME } from '@/actions/image-tool/constants';
@@ -416,16 +415,6 @@ export class WeaveImageNode extends WeaveNode {
         zIndex: 0,
       });
       this.updateImageCrop(nextProps);
-    }
-
-    try {
-      const selectionPlugin =
-        this.instance.getPlugin<WeaveNodesSelectionPlugin>('nodesSelection');
-      if (selectionPlugin) {
-        selectionPlugin.getTransformer().forceUpdate();
-      }
-    } catch (error) {
-      console.error('Error updating transformer', error);
     }
   }