+++ code/packages/sdk/src/plugins/copy-paste-nodes/types.ts
@@ -23,6 +23,7 @@ export type WeaveCopyPasteNodesPluginOnPasteEvent = Error | undefined;
 export type WeaveCopyPasteNodesPluginOnPasteExternalEvent = {
   items?: ClipboardItems;
   dataList?: DataTransferItemList;
+  positionCalculated: boolean;
   position: Vector2d;
 };